imrove error handling

+++ tasks/game/delete_replay.py
@@ -24,6 +24,11 @@ def delete_replay_folder(match_id: int, outer_logger=None, **kwargs):
         this_logger.info(f'Deleting match {match_id} replay folder')
 
         folder_path = Path(os.path.join(BASE_REPLAY_PATH, f'{match_id}'))
-        shutil.rmtree(folder_path)
+        if os.path.isdir(folder_path):
+            shutil.rmtree(folder_path)
+            this_logger.info(f"Deleted folder for game {match_id} successfully")
+        else:
+            this_logger.warning(f"No folder for game {match_id} was found")
+
 
     return None
